--- payment-example/src/main/java/com/verygoodsecurity/payment_example/data/repository/auth/AccessTokenRepository.kt
@@ -1,8 +1,0 @@
-package com.verygoodsecurity.payment_example.data.repository.auth
-
-import com.verygoodsecurity.payment_example.data.Result
-
-interface AccessTokenRepository {
-
-    suspend fun get(): Result<String>
-}
