--- payment-example/src/main/java/com/verygoodsecurity/payment_example/data/source/remote/RetrofitHelper.kt
@@ -1,23 +1,0 @@
-package com.verygoodsecurity.payment_example.data.source.remote
-
-import okhttp3.OkHttpClient
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import java.util.concurrent.TimeUnit
-
-internal object RetrofitHelper {
-
-    private const val TIMEOUT = 60L
-
-    fun build(url: String): Retrofit = Retrofit.Builder()
-        .baseUrl(url)
-        .client(createOkHttpClient())
-        .addConverterFactory(GsonConverterFactory.create())
-        .build()
-
-    private fun createOkHttpClient(): OkHttpClient = OkHttpClient.Builder()
-        .callTimeout(TIMEOUT, TimeUnit.SECONDS)
-        .readTimeout(TIMEOUT, TimeUnit.SECONDS)
-        .writeTimeout(TIMEOUT, TimeUnit.SECONDS)
-        .build()
-}
