--- payment-example/src/main/java/com/verygoodsecurity/payment_example/data/source/remote/auth/AccessTokenDataSource.kt
@@ -1,8 +1,0 @@
-package com.verygoodsecurity.payment_example.data.source.remote.auth
-
-import com.verygoodsecurity.payment_example.data.Result
-
-internal interface AccessTokenDataSource {
-
-    suspend fun get(): Result<String>
-}
