--- payment-example/src/main/java/com/verygoodsecurity/payment_example/data/source/remote/auth/dto/GetAccessTokenResponse.kt
@@ -1,7 +1,0 @@
-package com.verygoodsecurity.payment_example.data.source.remote.auth.dto
-
-import com.google.gson.annotations.SerializedName
-
-data class GetAccessTokenResponse constructor(
-    @SerializedName("access_token") val token: String
-)
