--- payment-example/src/main/java/com/verygoodsecurity/payment_example/data/source/remote/order/dto/CreateOrderResponse.kt
@@ -1,8 +1,0 @@
-package com.verygoodsecurity.payment_example.data.source.remote.order.dto
-
-import com.google.gson.annotations.SerializedName
-import com.verygoodsecurity.payment_example.data.entity.Order
-
-data class CreateOrderResponse constructor(
-    @SerializedName("data") val order: Order
-)
