+++ payment-example/src/main/java/com/verygoodsecurity/payments/data/Result.kt
@@ -1,0 +1,8 @@
+package com.verygoodsecurity.payments.data
+
+sealed class Result<out T : Any> {
+
+    data class Success<out T : Any>(val data: T) : Result<T>()
+
+    data class Error(val code: Int = -1, val msg: String? = null) : Result<Nothing>()
+}
