+++ payment-example/src/main/java/com/verygoodsecurity/payments/data/repository/auth/AccessTokenRepository.kt
@@ -1,0 +1,8 @@
+package com.verygoodsecurity.payments.data.repository.auth
+
+import com.verygoodsecurity.payments.data.Result
+
+interface AccessTokenRepository {
+
+    suspend fun get(): Result<String>
+}
