+++ payment-example/src/main/java/com/verygoodsecurity/payments/data/repository/transaction/TransactionRepository.kt
@@ -1,0 +1,14 @@
+package com.verygoodsecurity.payments.data.repository.transaction
+
+import com.verygoodsecurity.payments.data.Result
+import com.verygoodsecurity.payments.data.entity.Transaction
+
+interface TransactionRepository {
+
+    suspend fun create(
+        finId: String,
+        vaultId: String,
+        amount: Long,
+        currency: String
+    ): Result<Transaction>
+}
