+++ payment-example/src/main/java/com/verygoodsecurity/payments/data/source/remote/auth/AccessTokenDataSource.kt
@@ -1,0 +1,8 @@
+package com.verygoodsecurity.payments.data.source.remote.auth
+
+import com.verygoodsecurity.payments.data.Result
+
+internal interface AccessTokenDataSource {
+
+    suspend fun get(): Result<String>
+}
