+++ payment-example/src/main/java/com/verygoodsecurity/payments/data/source/remote/auth/dto/GetAccessTokenResponse.kt
@@ -1,0 +1,7 @@
+package com.verygoodsecurity.payments.data.source.remote.auth.dto
+
+import com.google.gson.annotations.SerializedName
+
+data class GetAccessTokenResponse constructor(
+    @SerializedName("access_token") val token: String
+)
