+++ payment-example/src/main/java/com/verygoodsecurity/payments/data/source/remote/order/OrderDataSource.kt
@@ -1,0 +1,9 @@
+package com.verygoodsecurity.payments.data.source.remote.order
+
+import com.verygoodsecurity.payments.data.Result
+import com.verygoodsecurity.payments.data.entity.Order
+
+internal interface OrderDataSource {
+
+    suspend fun create(): Result<Order>
+}
