+++ payment-example/src/main/java/com/verygoodsecurity/payments/data/source/remote/order/dto/CreateOrderResponse.kt
@@ -1,0 +1,8 @@
+package com.verygoodsecurity.payments.data.source.remote.order.dto
+
+import com.google.gson.annotations.SerializedName
+import com.verygoodsecurity.payments.data.entity.Order
+
+data class CreateOrderResponse constructor(
+    @SerializedName("data") val order: Order
+)
