+++ payment-example/src/main/java/com/verygoodsecurity/payments/data/source/remote/payment/dto/CreateTransactionRequest.kt
@@ -1,0 +1,10 @@
+package com.verygoodsecurity.payments.data.source.remote.payment.dto
+
+import com.google.gson.annotations.SerializedName
+
+data class CreateTransactionRequest constructor(
+    @SerializedName("fi_id") val finId: String,
+    @SerializedName("tnt") val vaultId: String,
+    @SerializedName("amount") val amount: Long,
+    @SerializedName("currency") val currency: String,
+)
